+++ lib/src/_scaffolding/configuration/endpoints.dart
@@ -3,6 +3,8 @@ library blackboard.rest.configuration.endpoints;
 /// The base endpoint for all REST API calls...
 const String base = '/learn/api/public/v1/';
 
+const String baseV2 = '/learn/api/public/v2';
+
 /// The oauth endpoints...
 const Map<String, String> oauth2 = {
   'authorization_code': 'oauth2/authorizationcode',