+++ lib/src/course_grades/anonymous_grading_type.dart
@@ -1,0 +1,4 @@
+/// The [AnonymousGradingType] enumeration...
+enum AnonymousGradingType {
+  None, AfterAllGraded, Date
+}