+++ lib/src/course_grades/attempt_status.dart
@@ -1,0 +1,5 @@
+/// The [AttemptStatus] enumeration...
+enum AttemptStatus {
+  NotAttempted, Abandoned, InProgress, Suspended, Canceled, NeedsGrading,
+  Completed, InMoreProgress, NeedsMoreGrading
+}