+++ lib/src/course_grades/grade_availability.dart
@@ -1,0 +1,4 @@
+/// The [GradeAvailability] enumeration...
+enum GradeAvailability {
+  Yes, No
+}