+++ lib/src/course_grades/grade_status.dart
@@ -1,0 +1,4 @@
+/// The [GradeStatus] enumeration...
+enum GradeStatus {
+  Graded, NeedsGrading
+}