+++ lib/src/course_grades/grading_type.dart
@@ -1,0 +1,4 @@
+/// The [GradingType] enumeration...
+enum GradingType {
+  Attempts, Calculated, Manual
+}