+++ lib/src/course_grades/scale_type.dart
@@ -1,0 +1,4 @@
+/// The [ScaleType] enumeration...
+enum ScaleType {
+  Percent, Score, Tabular, Text, Complete
+}