+++ lib/src/course_grades/score.dart
@@ -1,0 +1,9 @@
+library blackboard.rest.grades.score;
+
+/// The [Score] class...
+class Score {
+  final double possible;
+
+  /// The [Score] constructor...
+  Score (this.possible);
+}