+++ lib/src/course_grades/scoring_model.dart
@@ -1,0 +1,4 @@
+/// The [ScoringModel] enumeration...
+enum ScoringModel {
+  Last, Highest, Lowest, First, Average
+}